+++ src/main/java/lotto/model/machine/LottoMachine.java
@@ -1,4 +1,6 @@
-package lotto.model.lotto;
+package lotto.model.machine;
+
+import lotto.model.lotto.Lotto;
 
 public interface LottoMachine {
     Lotto generateLotto();
