+++ src/main/java/lotto/model/machine/RandomLottoMachine.java
@@ -1,4 +1,4 @@
-package lotto.model.lotto;
+package lotto.model.machine;
 
 import static lotto.utils.LottoConstant.LOTTO_SIZE;
 import static lotto.utils.LottoConstant.MAX_NUMBER;
@@ -6,8 +6,10 @@ import static lotto.utils.LottoConstant.MIN_NUMBER;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.List;
+import lotto.model.lotto.Lotto;
+import lotto.model.machine.LottoMachine;
 
-public class RandomLottoMachine implements LottoMachine{
+public class RandomLottoMachine implements LottoMachine {
     @Override
     public Lotto generateLotto() {
         List<Integer> numbers = generateRandomNumbers();
