+++ src/test/java/lotto/LottoTicketTest.java
@@ -5,7 +5,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import java.util.Arrays;
 import java.util.List;
 import lotto.model.Budget;
-import lotto.model.lotto.LottoMachine;
+import lotto.model.machine.LottoMachine;
 import lotto.model.lotto.LottoTicket;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
