+++ src/test/java/lotto/RandomLottoMachineTest.java
@@ -3,8 +3,8 @@ package lotto;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import lotto.model.lotto.Lotto;
-import lotto.model.lotto.LottoMachine;
-import lotto.model.lotto.RandomLottoMachine;
+import lotto.model.machine.LottoMachine;
+import lotto.model.machine.RandomLottoMachine;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
